Surface Traversal working
Unfortunately this doesn't work quite as well as I had at first planned.

I neglected to realize that a hard-coded locality check would result in
a traversal that always tends toward a local max/min.  This means that
the points found often converge, which is not really desirable as far as
I can tell.

It's possible that adding a heuristic which identifies the major plane
and traversing that plane first would yield a better result, however it
still seems like that would prove work only when there actually was a
major plane; ie. the surfaces AABB was not cubic